+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.8.0
+==================
+
+* Deprecate all funcs taking `seq[Rune]`
+
 v0.7.0
 ==================